chore(release): bump version to 0.23.0
- bridge: mTLS for localhost gRPC IPC (rc-tmtj)
- bridge: version bump 0.3.0 to 0.4.0
- bridge: rustls crypto provider in tests + CLI
- wasm: async source world with bidirectional streaming
- wasm: guest->host return-path streaming drain
- sec: batch8 config/dsl/log hardening
- sec: batch9 dataformats/eip DoS hardening
- sec: exec component fail-closed for v1.0
- http: TLS termination for consumer server
- grpc: bidirectional TLS + mTLS hardening
- kafka: ssl/ssl-vendored/sasl feature gates
- xtask: use runtime CWD for workspace root

Bd: rc-tmtj

+++ crates/components/camel-jms/src/component.rs
@@ -1829,8 +1829,8 @@ mod tests {
 
         let guard = pool.bridge_create_lock.lock().await;
         let total_count = pool.slots.len();
-        assert!(total_count >= pool.max_bridges as usize);
-        let result = if total_count >= pool.max_bridges as usize {
+        assert!(total_count >= pool.max_bridges);
+        let result = if total_count >= pool.max_bridges {
             Err(CamelError::Config(format!(
                 "JMS bridge limit reached: {total_count} bridge(s) >= max_bridges ({})",
                 pool.max_bridges
